♻️ #48 - 댓글 영역 추상화
- DiaryContainer, DiaryComments, DiaryComment, DiaryCommentInput 컴포넌트 분리
- 네이밍 컨벤션에 맞춰 수정
